Move `get_instance_func`
